feat(bitwarden): specify profile (#90)
As per the [bitwarden
documentation](https://bitwarden.com/help/cli/#log-in-to-multiple-accounts),
mulitple profiles are supported by specifying an environment variable
(`BITWARDENCLI_APPDATA_DIR `).

This code change sets that environment variable if a "profile" has been
specified in the bitwarden provider configuration.
 
We do not override the variable if that has already been set
(conservative approach).
 
The user is prompted for profile on `fnox init --provider bitwarden`.

---------

Co-authored-by: Ivo Bellin Salarin <ivo@nilleb.com>