feat(bitwarden): rbw support (experimental) (#91)
Now the bitwarden provider supports two backends (`bw` and `rbw` 🆕 ).
This was a popular discussion.

The init process assumes `bw` (no explicit question about this new
backend). If you wish to give it a try, add an attribute in the
corresponding provider configuration
```toml
[providers.bitwarden]
type = "bitwarden"
# this option ⬇️
backend = "rbw"
```

Disclaimer: the organizationID and collectionID are **not** supported by
the rbw backend; this should be considered experimental in the sense
that I have started using rbw today and I haven't tested anything beyond
the "note" support. Any feedback is welcome!

---------

Co-authored-by: Ivo Bellin Salarin <ivo@nilleb.com>