docs: add looping example for age provider (#106)
When going through the docs I felt that the need to re-encrypt every age
provider secret line by line cumbersome so I used a loop with `awk` and
figured it might be useful to include in the docs.

Of course, if you do not agree or want to include an example that relies
on other tools or is tightly coupled to the layout of `fnox list` then I
completely understand!

Just wanted add value to others if possible! Feel free to make any
changes or decline.

### Screenshot of the changes on the webpage
<img width="747" height="427" alt="image"
src="https://github.com/user-attachments/assets/cfedaea6-6a17-4224-8606-a0a63231f910"
/>